Bump Version 1.2.835

+++ builds/ansible/ddb-ansible.py
@@ -129,7 +129,7 @@ def run_module():
 # File   : ./source/ddb/version.py
 # ############################################################################
 
-__version__='1.2.834'
+__version__='1.2.835'
 
         
 # ############################################################################
@@ -3963,6 +3963,10 @@ def method_drop_table(context, meta):
     context.info("Drop Table")
     try:
         table=get_table(context,meta)
+        if table==None:
+            raise Exception("Table not found")
+        print table.data.name
+        print table.data.database
         results = context.database.drop_table(table_name=table.data.name,database_name=table.data.database)
         return query_results(success=results)
     except Exception as ex:
